test(fused_moe): consolidate warp decode JIT coverage
